Move vpn info proto conversion to nym-vpn-proto crate
