+++ nym-vpn-core/crates/nym-vpn-lib-types/Cargo.toml
@@ -19,6 +19,7 @@ nym-type-conversions = [
     "dep:nym-statistics-common",
     "dep:nym-vpn-api-client",
     "dep:nym-wg-gateway-client",
+    "dep:nym-vpn-network-config",
 ]
 
 [dependencies]
@@ -31,5 +32,6 @@ nym-connection-monitor = { workspace = true, optional = true }
 nym-gateway-directory = { workspace = true, optional = true }
 nym-statistics-common = { workspace = true, optional = true }
 nym-vpn-api-client = { workspace = true, optional = true }
+nym-vpn-network-config = { workspace = true, optional = true }
 nym-wg-gateway-client = { workspace = true, optional = true }
 strum_macros = { workspace = true }
