+++ nym-vpn-core/crates/nym-vpn-lib-types/src/service.rs
@@ -1,0 +1,16 @@
+// Copyright 2025 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: GPL-3.0-only
+
+use nym_vpn_network_config::{NymNetwork, NymVpnNetwork};
+use time::OffsetDateTime;
+
+#[derive(Clone, Debug)]
+pub struct VpnServiceInfo {
+    pub version: String,
+    pub build_timestamp: Option<OffsetDateTime>,
+    pub triple: String,
+    pub platform: String,
+    pub git_commit: String,
+    pub nym_network: NymNetwork,
+    pub nym_vpn_network: NymVpnNetwork,
+}
