+++ nym-vpn-core/crates/nym-vpn-proto/src/conversions/into_proto/service.rs
@@ -1,0 +1,30 @@
+// Copyright 2024 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: GPL-3.0-only
+
+use nym_vpn_lib_types::VpnServiceInfo;
+
+use crate::{
+    InfoResponse as ProtoInfoResponse, NymNetworkDetails as ProtoNymNetworkDetails,
+    NymVpnNetworkDetails as ProtoNymVpnNetworkDetails,
+};
+
+impl From<VpnServiceInfo> for ProtoInfoResponse {
+    fn from(info: VpnServiceInfo) -> Self {
+        let build_timestamp = info
+            .build_timestamp
+            .map(crate::conversions::prost::offset_datetime_into_proto_timestamp);
+
+        let nym_network = Some(ProtoNymNetworkDetails::from(info.nym_network.clone()));
+        let nym_vpn_network = Some(ProtoNymVpnNetworkDetails::from(info.nym_vpn_network));
+
+        Self {
+            version: info.version,
+            build_timestamp,
+            triple: info.triple,
+            platform: info.platform,
+            git_commit: info.git_commit,
+            nym_network,
+            nym_vpn_network,
+        }
+    }
+}
