+++ nym-vpn-core/crates/nym-vpnd/src/command_interface/connection_handler.rs
@@ -10,7 +10,7 @@ use nym_vpn_api_client::{
     NetworkCompatibility,
 };
 use nym_vpn_lib::gateway_directory::{EntryPoint, ExitPoint, GatewayClient, GatewayType};
-use nym_vpn_lib_types::TunnelState;
+use nym_vpn_lib_types::{TunnelState, VpnServiceInfo};
 use nym_vpn_network_config::{FeatureFlags, ParsedAccountLinks, SystemMessages};
 use nym_vpnd_types::gateway;
 
@@ -19,7 +19,6 @@ use crate::logging::LogPath;
 use crate::service::{
     AccountError, ConnectArgs, ConnectOptions, SetNetworkError, VpnServiceCommand,
     VpnServiceConnectError, VpnServiceDeleteLogFileError, VpnServiceDisconnectError,
-    VpnServiceInfo,
 };
 
 #[derive(Debug, thiserror::Error)]
